+++ tapo/src/api/hub_handler.rs
@@ -1,14 +1,11 @@
-use std::fmt;
+use std::sync::Arc;
 
-use serde::de::DeserializeOwned;
+use tokio::sync::RwLock;
 
 use crate::api::ApiClient;
 use crate::api::{KE100Handler, S200BHandler, T100Handler, T110Handler, T300Handler, T31XHandler};
 use crate::error::Error;
-use crate::requests::TapoRequest;
-use crate::responses::{
-    ChildDeviceHubResult, ChildDeviceListHubResult, DeviceInfoHubResult, TapoResponseExt,
-};
+use crate::responses::{ChildDeviceHubResult, ChildDeviceListHubResult, DeviceInfoHubResult};
 
 macro_rules! get_device_id {
     ($self:expr, $identifier:expr, $value:path) => {{
@@ -37,18 +34,20 @@ macro_rules! get_device_id {
 
 /// Handler for the [H100](https://www.tapo.com/en/search/?q=H100) hubs.
 pub struct HubHandler {
-    client: ApiClient,
+    client: Arc<RwLock<ApiClient>>,
 }
 
 /// Hub handler methods.
 impl HubHandler {
     pub(crate) fn new(client: ApiClient) -> Self {
-        Self { client }
+        Self {
+            client: Arc::new(RwLock::new(client)),
+        }
     }
 
     /// Refreshes the authentication session.
     pub async fn refresh_session(&mut self) -> Result<&mut Self, Error> {
-        self.client.refresh_session().await?;
+        self.client.write().await.refresh_session().await?;
         Ok(self)
     }
 
@@ -56,13 +55,13 @@ impl HubHandler {
     /// It is not guaranteed to contain all the properties returned from the Tapo API.
     /// If the deserialization fails, or if a property that you care about it's not present, try [`HubHandler::get_device_info_json`].
     pub async fn get_device_info(&self) -> Result<DeviceInfoHubResult, Error> {
-        self.client.get_device_info().await
+        self.client.read().await.get_device_info().await
     }
 
     /// Returns *device info* as [`serde_json::Value`].
     /// It contains all the properties returned from the Tapo API.
     pub async fn get_device_info_json(&self) -> Result<serde_json::Value, Error> {
-        self.client.get_device_info().await
+        self.client.read().await.get_device_info().await
     }
 
     /// Returns *child device list* as [`ChildDeviceHubResult`].
@@ -71,6 +70,8 @@ impl HubHandler {
     /// If the deserialization fails, or if a property that you care about it's not present, try [`HubHandler::get_child_device_list_json`].
     pub async fn get_child_device_list(&self) -> Result<Vec<ChildDeviceHubResult>, Error> {
         self.client
+            .read()
+            .await
             .get_child_device_list::<ChildDeviceListHubResult>()
             .await
             .map(|r| r.devices)
@@ -79,25 +80,17 @@ impl HubHandler {
     /// Returns *child device list* as [`serde_json::Value`].
     /// It contains all the properties returned from the Tapo API.
     pub async fn get_child_device_list_json(&self) -> Result<serde_json::Value, Error> {
-        self.client.get_child_device_list().await
+        self.client.read().await.get_child_device_list().await
     }
 
     /// Returns *child device component list* as [`serde_json::Value`].
     /// This information is useful in debugging or when investigating new functionality to add.
     pub async fn get_child_device_component_list_json(&self) -> Result<serde_json::Value, Error> {
-        self.client.get_child_device_component_list().await
-    }
-
-    /// Internal method that's called by functions of the child devices.
-    pub(crate) async fn control_child<R>(
-        &self,
-        device_id: String,
-        request_data: TapoRequest,
-    ) -> Result<Option<R>, Error>
-    where
-        R: fmt::Debug + DeserializeOwned + TapoResponseExt,
-    {
-        self.client.control_child(device_id, request_data).await
+        self.client
+            .read()
+            .await
+            .get_child_device_component_list()
+            .await
     }
 }
 
@@ -129,7 +122,7 @@ impl HubHandler {
     /// ```
     pub async fn ke100(&self, identifier: HubDevice) -> Result<KE100Handler, Error> {
         let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::KE100);
-        Ok(KE100Handler::new(self, device_id))
+        Ok(KE100Handler::new(self.client.clone(), device_id))
     }
 
     /// Returns a [`S200BHandler`] for the given [`HubDevice`].
@@ -158,7 +151,7 @@ impl HubHandler {
     /// ```
     pub async fn s200b(&self, identifier: HubDevice) -> Result<S200BHandler, Error> {
         let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::S200B);
-        Ok(S200BHandler::new(self, device_id))
+        Ok(S200BHandler::new(self.client.clone(), device_id))
     }
 
     /// Returns a [`T100Handler`] for the given [`HubDevice`].
@@ -187,7 +180,7 @@ impl HubHandler {
     /// ```
     pub async fn t100(&self, identifier: HubDevice) -> Result<T100Handler, Error> {
         let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::T100);
-        Ok(T100Handler::new(self, device_id))
+        Ok(T100Handler::new(self.client.clone(), device_id))
     }
 
     /// Returns a [`T110Handler`] for the given [`HubDevice`].
@@ -216,7 +209,7 @@ impl HubHandler {
     /// ```
     pub async fn t110(&self, identifier: HubDevice) -> Result<T110Handler, Error> {
         let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::T110);
-        Ok(T110Handler::new(self, device_id))
+        Ok(T110Handler::new(self.client.clone(), device_id))
     }
 
     /// Returns a [`T300Handler`] for the given [`HubDevice`].
@@ -245,7 +238,7 @@ impl HubHandler {
     /// ```
     pub async fn t300(&self, identifier: HubDevice) -> Result<T300Handler, Error> {
         let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::T300);
-        Ok(T300Handler::new(self, device_id))
+        Ok(T300Handler::new(self.client.clone(), device_id))
     }
 
     /// Returns a [`T31XHandler`] for the given [`HubDevice`].
@@ -274,7 +267,7 @@ impl HubHandler {
     /// ```
     pub async fn t310(&self, identifier: HubDevice) -> Result<T31XHandler, Error> {
         let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::T310);
-        Ok(T31XHandler::new(self, device_id))
+        Ok(T31XHandler::new(self.client.clone(), device_id))
     }
 
     /// Returns a [`T31XHandler`] for the given [`HubDevice`].
@@ -303,7 +296,7 @@ impl HubHandler {
     /// ```
     pub async fn t315(&self, identifier: HubDevice) -> Result<T31XHandler, Error> {
         let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::T315);
-        Ok(T31XHandler::new(self, device_id))
+        Ok(T31XHandler::new(self.client.clone(), device_id))
     }
 }